requirements.txt
django>=4.2,<5.1
pillow>=10.0
